Parallax and polishing
-Added Parallax effect (Thanks @MEEPofFaith) to Pyroclast

-More polishing for pyroclast

-Added in failsafe for when other mods change Frostscape bullet data

-Moved BouncyBulletType into bullet sub-directory inside entities

...I swear I don't want to deal with another mod which thinks changing other bullet's data is a good idea.

Changing bullet data outside a BulletType is a bad idea due to more complex mods tending to use data within their BulletType. If it's changed outside a BulletType to an unintended value, expect chaos. It's not their fault for building their bullet to work this way with data intended to be x in an environment where almost no mods change bullet data outside of BulletType, and you change it to y

(Im tired. Im sorry about the rant, it had to go somewhere)